refactor(backend): utils 파일을 도메인 모듈로 이동
- wsRawData.ts → modules/claude/ (claude 모듈에서만 사용)
- directoryParser.ts → modules/files/ (files 모듈에서만 사용)
- batchAsync.ts 삭제 (프로덕션 사용처 없음)
- 빈 utils/ 폴더 제거
